refactoring without type of blog
