+++ src/pages/_document.js
@@ -31,7 +31,7 @@ class MyDocument extends Document {
 
     render() {
         return (
-            <html lang="en" dir="ltr">
+            <html lang="jp" dir="ltr">
                 <Head>
                     <meta charSet="utf-8" />
                     {/* Use minimum-scale=1 to enable GPU rasterization */}
