+++ src/pages/blogs/[id].js
@@ -1,11 +1,8 @@
-import React from 'react'
+import React, { Fragment } from 'react'
 import fetch from 'isomorphic-unfetch'
-import HeaderList from '../../components/header/HeaderList'
-import Header from '../../components/header/Header'
-import Footer from '../../components/Footer'
-import SwipeDrawer from '../../components/header/SwipeDrawer'
+import { Template } from '../../components/common/Template'
 import { makeStyles } from '@material-ui/core/styles'
-import { Container, Grid, Hidden, Typography } from '@material-ui/core'
+import { Container, Grid, Typography } from '@material-ui/core'
 
 export default function blogId({ blog }) {
     const classes = useStyles()
@@ -14,86 +11,68 @@ export default function blogId({ blog }) {
     )
     return (
         <>
-            <Header>
-                <Hidden smUp>
-                    <SwipeDrawer />
-                </Hidden>
-                <Hidden xsDown>
-                    <HeaderList />
-                </Hidden>
-            </Header>
-            <main>
-                <div className={classes.heroContent}>
-                    <Container>
-                        <Grid
-                            container
-                            alignItems="center"
-                            justify="center"
-                            className={classes.mainContent}
-                        >
-                            <Grid className={classes.mainContentInner}>
-                                <Grid
-                                    container
-                                    alignItems="center"
-                                    justify="center"
-                                >
-                                    {typeof blog.image === 'undefined' ? (
-                                        ''
-                                    ) : (
-                                        <Image image={blog.image} />
-                                    )}
-                                </Grid>
-                                <div>
-                                    <Typography
-                                        variant={'h2'}
-                                        align="center"
-                                        className={classes.title}
-                                    >
-                                        {blog.title}
-                                    </Typography>
-                                    <div>
-                                        {blog.tags.map((tag) => (
-                                            <React.Fragment key={tag.id}>
-                                                <Typography
-                                                    variant={'h6'}
-                                                    align="right"
-                                                    className={classes.tag}
-                                                >
-                                                    {tag.name}
-                                                </Typography>
-                                            </React.Fragment>
-                                        ))}
-                                    </div>
-                                    <Typography
-                                        variant={'h4'}
-                                        align="center"
-                                        className={classes.body}
-                                    >
-                                        <div
-                                            dangerouslySetInnerHTML={{
-                                                __html: `${blog.body}`,
-                                            }}
-                                        ></div>
-                                    </Typography>
-                                </div>
+            <Template color="linear-gradient(to left, #efefbb, #d4d3dd)">
+                <Container>
+                    <Grid
+                        container
+                        alignItems="center"
+                        justify="center"
+                        className={classes.mainContent}
+                    >
+                        <Grid className={classes.mainContentInner}>
+                            <Grid
+                                container
+                                alignItems="center"
+                                justify="center"
+                            >
+                                {typeof blog.image === 'undefined' ? (
+                                    ''
+                                ) : (
+                                    <Image image={blog.image} />
+                                )}
                             </Grid>
+                            <div>
+                                <Typography
+                                    variant={'h2'}
+                                    align="center"
+                                    className={classes.title}
+                                >
+                                    {blog.title}
+                                </Typography>
+                                <div>
+                                    {blog.tags.map((tag) => (
+                                        <Fragment key={tag.id}>
+                                            <Typography
+                                                variant={'h6'}
+                                                align="right"
+                                                className={classes.tag}
+                                            >
+                                                {tag.name}
+                                            </Typography>
+                                        </Fragment>
+                                    ))}
+                                </div>
+                                <Typography
+                                    variant={'h4'}
+                                    align="center"
+                                    className={classes.body}
+                                >
+                                    <div
+                                        dangerouslySetInnerHTML={{
+                                            __html: `${blog.body}`,
+                                        }}
+                                    ></div>
+                                </Typography>
+                            </div>
                         </Grid>
-                    </Container>
-                </div>
-            </main>
-            <Footer description="Hair and Web designer" />
+                    </Grid>
+                </Container>
+            </Template>
         </>
     )
 }
 
-const useStyles = makeStyles((theme) => ({
-    heroContent: {
-        padding: theme.spacing(8, 0, 6),
-        paddingTop: '8%',
-        width: '100%',
-        background: 'linear-gradient(to left, #efefbb, #d4d3dd)',
-        overflow: 'hidden',
-    },
+const useStyles = makeStyles({
     mainContent: {
         backgroundColor: '#f8f8f8',
         width: '100%',
@@ -124,7 +103,7 @@ const useStyles = makeStyles((theme) => ({
         width: '30%',
         marginTop: '3em',
     },
-}))
+})
 
 export const getStaticPaths = async () => {
     const key = {
