+++ src/pages/work.js
@@ -1,27 +1,22 @@
-import React from 'react'
-import HeaderList from '../components/header/HeaderList'
-import Header from '../components/header/Header'
-import Footer from '../components/Footer'
-import SwipeDrawer from '../components/header/SwipeDrawer'
-import LottieCamera from '../components/Lottie/camera/camera'
-import LottiePc from '../components/Lottie/pcScreen/pcScreen'
-import workLogo from '../../public/static/img/WORK.png'
-import { Container, Grid, Hidden, Button } from '@material-ui/core'
+import React, { useEffect, useState } from 'react'
+import { Container, Grid } from '@material-ui/core'
 import { makeStyles } from '@material-ui/core/styles'
-import LottieNoteAndPen from '../components/Lottie/noteAndPen/noteAndPen'
 import fetch from 'isomorphic-unfetch'
+//components
 import { ImageAnimation } from '../components/helper/animationHelper'
+import LottieButton from '../components/work/LottieButton'
+import { WorkLogo } from '../components/common/topLogoSection'
+import { Template } from '../components/common/Template'
 
 export default function work({ images }) {
-    const [switchImages, setSwitchImages] = React.useState(0)
-
-    React.useEffect(() => {
-        ImageAnimation('.line', '.image', 0.95, 0)
-    }, [switchImages])
+    const [switchImages, setSwitchImages] = useState(0)
 
     const onClickHandler = (num) => (event) => {
         setSwitchImages(num)
     }
+    useEffect(() => {
+        ImageAnimation('.line', '.image', 0.95, 0)
+    }, [switchImages])
 
     const hair = images.filter((hair) => {
         return hair.imageTags[0].tags == 'Black_and_White' && !undefined
@@ -36,112 +31,50 @@ export default function work({ images }) {
     const classes = useStyles()
     return (
         <>
-            <Header>
-                <Hidden smUp>
-                    <SwipeDrawer />
-                </Hidden>
-                <Hidden xsDown>
-                    <HeaderList />
-                </Hidden>
-            </Header>
-            <main>
-                <div className={classes.heroContent}>
-                    <Grid className={classes.gap}>
-                        <Grid>
-                            <img
-                                className={classes.imgLogo}
-                                src={workLogo}
-                                alt="work_logo"
-                            />
-                            <hr />
+            <Template color="linear-gradient(to right, #ddd6f3, #faaca8)">
+                <WorkLogo />
+                <Container>
+                    <Grid className={classes.mainContent}>
+                        <LottieButton onClickHandler={onClickHandler} />
+                        <Grid
+                            container
+                            justify="space-evenly"
+                            direction="row"
+                            className={`${classes.images} imageContainer`}
+                        >
+                            {imageStore[switchImages].map((img) => (
+                                <div
+                                    style={{
+                                        position: 'relative',
+                                        width: '300px',
+                                        height: '320px',
+                                        overflow: 'hidden',
+                                    }}
+                                    key={img.id}
+                                >
+                                    <span
+                                        className={`${classes.line} line`}
+                                    ></span>
+                                    <img
+                                        className={`${classes.image} image`}
+                                        src={`${img.image.url}?fit=crop&${
+                                            img.imageTags[1]
+                                                ? 'w=600&h=400'
+                                                : 'w=300&h=300'
+                                        }`}
+                                        alt={img.id}
+                                    />
+                                </div>
+                            ))}
                         </Grid>
                     </Grid>
-                    <Container>
-                        <Grid className={classes.mainContent}>
-                            <div style={{ paddingTop: '50px' }}>
-                                <Grid
-                                    container
-                                    direction="row"
-                                    justify="space-evenly"
-                                    alignItems="center"
-                                >
-                                    <Button
-                                        variant="outlined"
-                                        color="default"
-                                        startIcon={<LottieCamera />}
-                                        className={classes.button}
-                                        onClick={onClickHandler(0)}
-                                    >
-                                        Hair
-                                    </Button>
-                                    <Button
-                                        variant="outlined"
-                                        color="primary"
-                                        startIcon={<LottiePc />}
-                                        className={classes.button}
-                                        onClick={onClickHandler(1)}
-                                    >
-                                        Web
-                                    </Button>
-                                    <Button
-                                        variant="outlined"
-                                        color="secondary"
-                                        startIcon={<LottieNoteAndPen />}
-                                        className={classes.button}
-                                        onClick={onClickHandler(2)}
-                                    >
-                                        Other
-                                    </Button>
-                                </Grid>
-                            </div>
-                            <Grid
-                                container
-                                justify="space-evenly"
-                                direction="row"
-                                className={`${classes.images} imageContainer`}
-                            >
-                                {imageStore[switchImages].map((img) => (
-                                    <div
-                                        style={{
-                                            position: 'relative',
-                                            width: '300px',
-                                            height: '320px',
-                                            overflow: 'hidden',
-                                        }}
-                                        key={img.id}
-                                    >
-                                        <span
-                                            className={`${classes.line} line`}
-                                        ></span>
-                                        <img
-                                            className={`${classes.image} image`}
-                                            src={`${img.image.url}?fit=crop&${
-                                                img.imageTags[1]
-                                                    ? 'w=600&h=400'
-                                                    : 'w=300&h=300'
-                                            }`}
-                                            alt={img.id}
-                                        />
-                                    </div>
-                                ))}
-                            </Grid>
-                        </Grid>
-                    </Container>
-                </div>
-            </main>
-            <Footer description="Hair and Web designer" />
+                </Container>
+            </Template>
         </>
     )
 }
 
 const useStyles = makeStyles((theme) => ({
-    heroContent: {
-        padding: theme.spacing(8, 0, 6),
-        minHeight: '800px',
-        width: '100%',
-        background: 'linear-gradient(to right, #ddd6f3, #faaca8)',
-        overflow: 'hidden',
-    },
     mainContent: {
         backgroundColor: '#f8f8f8',
         width: '100%',
@@ -149,47 +82,6 @@ const useStyles = makeStyles((theme) => ({
         minHeight: '1000px',
         borderRadius: '20px',
         boxShadow: '1px 1px 5px #fff',
-    },
-    gap: {
-        width: '100%',
-        height: '13rem',
-        display: 'table',
-        textAlign: 'center',
-        '@media (max-width:600px)': {
-            height: '8rem',
-        },
-        '& div': {
-            display: 'table-cell',
-            verticalAlign: 'middle',
-            '& img': {
-                zIndex: 1,
-                position: 'relative',
-                opacity: 0.7,
-                animation:
-                    '$fade-in-top 1s cubic-bezier(0.390, 0.575, 0.565, 1.000) both',
-            },
-            '& hr': {
-                width: '100%',
-                height: '1.5em',
-                backgroundColor: '#fff',
-                borderWidth: 'inherit',
-                borderStyle: 'none',
-                transform: 'translateY(-60px)',
-                zIndex: 0,
-                opacity: 0.7,
-                '@media (max-width:600px)': {
-                    height: '1em',
-                    transform: 'translateY(-50px)',
-                },
-            },
-        },
-    },
-    imgLogo: {
-        width: '14em',
-        height: 'auto',
-        '@media (max-width:600px)': {
-            width: '11em',
-        },
     },
     images: {
         width: '100%',
@@ -211,25 +103,6 @@ const useStyles = makeStyles((theme) => ({
         backgroundColor: '#9e9e9e',
         transformOrigin: '0 50%',
         zIndex: 100,
-    },
-    button: {
-        borderRadius: '30px',
-        border: '2px solid',
-        fontSize: '1em',
-        '@media (max-width:600px)': {
-            fontSize: '0.875em',
-            transform: 'scale(0.8)',
-        },
-    },
-    '@keyframes fade-in-top': {
-        '0%': {
-            transform: 'translateY(-30px)',
-            opacity: 0,
-        },
-        '100%': {
-            transform: 'translateY(0)',
-            opacity: 0.7,
-        },
     },
 }))
